--- OneCheck/Source/CustomView/Marker.swift
@@ -1,26 +1,0 @@
-//
-//  Marker.swift
-//  OneCheck
-//
-//  Created by 김동우 on 2023/05/26.
-//
-
-import MapKit
-
-class Marker: NSObject, MKAnnotation {
-  let title: String?
-  let coordinate: CLLocationCoordinate2D
-  let subtitle:String?
-
-  init(
-    title: String?,
-    subtitle: String?,
-    coordinate: CLLocationCoordinate2D
-  ) {
-    self.title = title
-    self.subtitle = subtitle
-    self.coordinate = coordinate
-
-    super.init()
-  }
-}
